Move user_service into user_api/api
